chore(vats): change makeStakeReporter to bootstrap behavior
