+++ packages/vats/src/core/types.js
@@ -178,6 +178,7 @@
  *     getRUNBundle: ERef<SourceBundle>,
  *     governanceBundles: ERef<typeof import('@agoric/run-protocol/src/importedBundles.js').governanceBundles>,
  *     initialSupply: ERef<Payment>,
+ *     lienBridge: ERef<StakingAuthority>,
  *     namesByAddress: ERef<NameHub>,
  *     namesByAddressAdmin: ERef<NameAdmin>,
  *     pegasusBundle: Promise<SourceBundle>,
@@ -201,6 +202,7 @@
  *     vaultBundles: Producer<Record<'VaultFactory' | 'liquidate', SourceBundle>>,
  *     governanceBundles: Producer<typeof import('@agoric/run-protocol/src/importedBundles.js').governanceBundles>,
  *     initialSupply: Producer<Payment>,
+ *     lienBridge: Producer<StakingAuthority>,
  *     centralSupplyBundle: Producer<SourceBundle>,
  *     feeMintAccess: Producer<FeeMintAccess>,
  *     priceAuthorityVat: Producer<PriceAuthorityVat>,
